Added information to Cold Brew, Drip Machine and French Press. Refactored method view to also accomodate french press method. Closes #8.
