+++ Bean Juice/Method Information/Chemex.md
@@ -1,6 +1,7 @@
 The Chemex® coffeemaker is an elegant, one-piece, hourglass shaped vessel made of high quality, heat resistant glass. The traditional model comes to you with a polished wood collar and leather tie. The collar serves as an insulated handle around the middle of the coffeemaker.
 
 ## History
+
 The CHEMEX® coffeemaker was invented in 1941 by the eccentric chemist Dr. Peter Schlumbohm. Throughout his career he developed over 300 patents, ranging from cocktail shakers to automobiles. His focus was on making everyday objects more functional, attractive and enjoyable to use.
 
 When designing the CHEMEX®, Schlumbohm desired to not only make brewing the perfect cup simple, but also to have the vessel be a thing of beauty. Being a chemist, he studied and understood clearly the chemistry behind the extraction of flavor and caffeine from coffee beans.
@@ -8,6 +9,7 @@ When designing the CHEMEX®, Schlumbohm desired to not only make brewing the per
 It was this knowledge that led him to invent the double bonded CHEMEX® paper filters for a perfect extraction every time. For the carafe, Schlumbohm was inspired by the Bauhaus school of design and non-porous labware that would impart no flavor of its own. Using these elements he fashioned the hourglass shape that has now become an iconic part of American history. Made simply from one piece of borosilicate glass and adorned with a wooden collar and rawhide tie, the CHEMEX® is pure in both form and function. It has been recognized and awarded by both the scientific community as well as the art and design communities. The CHEMEX® can be found in museums throughout the world and is included in the permanent collection at the Brooklyn Museum, Corning Museum of Glass and MOMA, NY.
 
 ## Usage
+
 1. Select your favorite whole bean coffee you want to enjoy and grind it to medium coarse ground.
 
 2. Open a CHEMEX® Bonded Coffee Filter into a cone shape so that one side of the cone has three layers, and place it into the top of the CHEMEX® brewer. The thick (three-layer) portion should cover the pouring spout.
