+++ Bean Juice/Method Information/French Press.md
@@ -1,0 +1,22 @@
+A French press, also known as a cafetière, cafetière à piston, caffettiera a stantuffo, press pot, coffee press, or coffee plunger, is a coffee brewing device, although it can also be used for other tasks. In 1923 Ugo Paolini, an Italian, lodged patent documents relating to a tomato juice separator and he developed the idea of making a coffee pot with a press action and a filter. He assigned his 1928 patent to Italian designer Attilio Calimani and Giulio Moneta who filed it in 1929.
+In English, the device is known in North America as a French press or coffee press; in Britain and Ireland as a cafetière; in New Zealand, Australia, and South Africa, as a coffee plunger, and coffee brewed in it as plunger coffee. In Italian it is known as a caffettiera a stantuffo; in German, as a Stempelkanne ("stamp pot") or Kaffeepresse ("coffee press"); in French as cafetière à piston, or simply as cafetière (also the usage in Dutch), though some speakers might also use genericized trademarks, such as Melior or Bodum.
+
+## History
+
+Over the years, the French press has undergone several design modifications. The first coffee press, which may have been made in France, was the modern coffee press in its rudimentary form—a metal or cheesecloth screen fitted to a rod that users would press into a pot of hot water and coffee grounds. Two French inventors (Mayer and Delforge) patented in 1852 a forerunner of the French press. A patent was filed by a Frenchman, Marcel-Pierre Paquet dit Jolbert, officially published on August 5, 1924.
+
+A coffee press was patented in the United States by Milanese designer Attilio Calimani in 1929. It underwent several design modifications through Faliero Bondanini, who patented his own version in 1958 and began manufacturing it in a French clarinet factory called Martin SA under the brand name "Melior". Its popularity may have been aided in 1965 by its use in the Michael Caine film The Ipcress File. The device was further popularized across Europe by a British company by the name of Household Articles Ltd. and the Danish tableware and kitchenware company Bodum.
+
+## Design
+
+The modern French press consists of a narrow cylindrical beaker, usually made of glass or clear plastic, equipped with a metal or plastic lid and plunger that fits tightly in the cylinder and has a fine stainless steel wire or nylon mesh filter.
+
+## Usage
+
+Coffee is brewed by placing coarsely ground coffee in the empty beaker and adding hot—between 93–96 °C (199–205 °F)—water, in proportions of about 30 g (1.1 oz) of coffee grounds to 500 ml (17 US fl oz) of water, more or less to taste. It is sometimes recommended that the grounds be pre-infused with a small amount of hot water. Plunging slowly is purported to maximize the extraction of the oils and flavonoids from the ground bean. The mesh piston normally does not compress the coffee grounds, as most designs leave a generous space—about 30 mm (1.2 in)—below the piston in its lowest position. If the brewed coffee is allowed to remain in the beaker with the used grounds, the coffee may become astringent and bitter, though this is an effect that some users of the French press consider desirable.
+
+A French press works best with coffee of a coarser grind than does a drip brew coffee filter, about the consistency of cooking salt. Finer coffee grounds, when immersed in water, have lower permeability, requiring an excessive amount of force to be applied by hand to lower the plunger and are more likely to seep through or around the perimeter of the press filter and into the coffee drink. Additionally, finer grounds will tend to over-extract and cause the coffee to taste bitter.
+
+Some writers give the optimum time for brewing as around four minutes. Other approaches, such as cold brewing, require several hours of contact between the water and the grounds to achieve the desired extraction.
+
+_Source: [Wikipedia](https://en.wikipedia.org/wiki/French_press)_
